+++ envsetup.sh
@@ -3,3 +3,7 @@
 st_build() {
 	go build -o sndtool ./cmd/
 }
+
+st_format() {
+	gofmt -w .
+}